requirements.txt
streamlit==1.31.0
groq==0.4.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
